pack rpcs

+++ Modules/Roles/Neutrals/AmnesiacRole.cs
@@ -52,7 +52,7 @@ public class AmnesiacRole : CustomRoleBehavior
                             {
                                 if (_player.IsLocalPlayer()) Utils.FlashScreen(RoleColor);
 
-                                _player.SendRpcSetCustomRole((int)data.BetterData().RoleInfo.RoleTypeWhenAlive);
+                                _player.SendRpcSetCustomRole(data.BetterData().RoleInfo.RoleTypeWhenAlive);
 
                                 /*
                                 _player.ClearAddonsSync();
